+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 <!-- Keep this up to date! After a release, change the tag name to the latest release -->
 [unreleased changes details]: https://github.com/Adobe-Consulting-Services/acs-aem-commons/compare/acs-aem-commons-4.7.2...HEAD
 
+### Added
+- #2434 - New workflow process step "Set Image Orientation"
+
 ## 4.8.6 - 2020-10-13
 
 ### Fixed
@@ -21,7 +24,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 ### Added
 - #1060 - New tree activation MCP utility
-- #New workflow process step "Set Image Orientation"
 
 ### Changed